US137422 - Commonize IPSIS paging methods
